requirements.txt
python-docx
PyYAML
pytest
matplotlib
